requirements.txt
Django==1.8.5
django-image-cropping==1.0.2
easy-thumbnails==2.2